Move rate limiting check
Noticed that the rate limiting check is in the middle of a loop that does not interact with Github APIs.

So, it is making the `gh api -i "repos/$gh_org" 2>&1 | grep -Ei '^X-Ratelimit-Remaining'` every user, regardless of it we will need to interact with the Github API for that user. 

By moving it, we can skip that check for all users who do use Copilot or were recently added.